+++ CHANGELOG.md
@@ -3,4 +3,7 @@ All notable changes to this project will be documented in this file.
  
 ## [Unreleased]
 
+### Added
+* Add micro_service PerunIdentity
+
 [Unreleased]: https://github.com/CESNET/satosa-module-perun/tree/master